migrate

+++ abc-laravel8-master/database/migrations/2021_12_22_074653_them_thong_tin_xac_admin_customer_manager_to_user.php
@@ -15,7 +15,7 @@ class ThemThongTinXacAdminCustomerManagerToUser extends Migration
     {
         Schema::table('users', function (Blueprint $table) {
             //->comment('0: for Admin and 1: for user or customer');
-            $table->tinyInteger('level')->after('password')->default(0)->comment('0: cho admin hay manager và 1: là khách hàng ')->change();
+            $table->smallInteger('level')->after('password')->default(0)->comment('0: cho admin hay manager và 1: là khách hàng ')->change();
         });
     }
